backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
branca==0.8.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filetype==1.2.0
folium==0.18.0
fonttools==4.55.5
future==1.0.0
geocoder==1.38.1
git-filter-repo==2.47.0
h11==0.16.0
idna==3.7
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mysql-connector-python==9.4.0
nltk==3.8.1
numpy==1.23.4
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==1.5.1
phonenumbers==8.13.50
pillow==11.1.0
pycparser==2.22
pygame==2.4.0
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.5
PyYAML==6.0.2
ratelim==0.1.6
regex==2023.10.3
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.51
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.34.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tabulate==0.9.0
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
xyzservices==2024.9.0
